brain/requirements.txt
openai
requests
python-dotenv
streamlit
fastapi
uvicorn
pydantic